summarization/pyproject.toml
[project]
name = "summarization"
version = "0.1.0"
description = ""
authors = [
    {name = "Ahmad Akra"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "transformers (>=4.51.3,<5.0.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "protobuf (>=6.30.2,<7.0.0)",
    "bson (>=0.5.10,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
